re-arrange tests for template functions
