+++ __tests__/unit/templates/group.test.js
@@ -1,5 +1,7 @@
 const { createGroupsTemplateForService, RESOLVERS } = require("../../../lib/templates/group");
 const defaults = require("../../../lib/defaults");
+const cds = require("@sap/cds");
+const { ORD_ACCESS_STRATEGY } = require("../../../lib/constants");
 
 const BASE_SERVICE = {
     name: "sap.test.BusinessPartnerService",
@@ -57,13 +59,34 @@ describe("RESOLVERS.groupId", () => {
     it("uses internalNamespace prefix when it matches service name", () => {
         const appConfig = { ...BASE_APP_CONFIG, internalNamespace: "sap.internal" };
         const service = { name: "sap.internal.MyService" };
-        expect(RESOLVERS.groupId(service, appConfig)).toBe(
-            `${defaults.groupTypeId}:sap.test:MyService`,
-        );
+        expect(RESOLVERS.groupId(service, appConfig)).toBe(`${defaults.groupTypeId}:sap.test:MyService`);
     });
 });
 
 describe("createGroupsTemplateForService", () => {
+    let serviceDefinition;
+    const appConfig = {
+        ordNamespace: "customer.testNamespace",
+        appName: "testAppName",
+        lastUpdate: "2022-12-19T15:47:04+00:00",
+        policyLevels: ["none"],
+        authConfig: {
+            accessStrategies: [ORD_ACCESS_STRATEGY.Open],
+        },
+    };
+
+    beforeAll(() => {
+        const model = cds.linked(`
+                service testServiceName {
+                    entity Books {
+                        key ID: UUID;
+                        title: String;
+                    }
+                };
+            `);
+        serviceDefinition = model.definitions["testServiceName"];
+    });
+
     it("produces a complete group object with defaults", () => {
         const result = createGroupsTemplateForService(BASE_SERVICE, BASE_APP_CONFIG);
         expect(result).toEqual({
@@ -82,5 +105,21 @@ describe("createGroupsTemplateForService", () => {
     it("always uses the canonical groupTypeId from defaults", () => {
         const result = createGroupsTemplateForService(BASE_SERVICE, BASE_APP_CONFIG);
         expect(result.groupTypeId).toBe(defaults.groupTypeId);
+    });
+
+    it("should return default value when groupIds do not have groupId", () => {
+        expect(createGroupsTemplateForService(serviceDefinition, appConfig)).toEqual({
+            groupId: "sap.cds:service:customer.testNamespace:testServiceName",
+            groupTypeId: "sap.cds:service",
+            title: "test Service",
+        });
+    });
+
+    it('should return default value with a proper Service title when "Service" keyword is missing', () => {
+        expect(createGroupsTemplateForService({ ...serviceDefinition, name: "testServName" }, appConfig)).toEqual({
+            groupId: "sap.cds:service:customer.testNamespace:testServName",
+            groupTypeId: "sap.cds:service",
+            title: "testServName Service",
+        });
     });
 });
